Add files via upload
This is a Java program that calculates the area of a rectangle based on user input. It prompts the user to enter the width and height of the rectangle using a Scanner object, calculates the area by multiplying the width and height, and then displays the result to the user.